refactor(provider): separate context actions and state into custom hook in order to unit test state
